feat: interactive debug builder
Problem: when large builds (spack) fail, we have to start over.
Solution: prototype a debug solution that, when a flag -i
(for interactive) is added for a custom build entrypoint, we
allow a dispatchRun builder command to fail, and still commit
the layer. This means we have the changes regardless of the exit
code. Note that this is implemented for the Dockerfile builder
but not yet buildkit, which has another grpc call that I have not
looked at yet. It should be considered a proof of concept.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>